+++ R/dublincore.R
@@ -19,7 +19,7 @@
 #' @details The \code{ResourceType} property will be by definition "Dataset".
 #' The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to the dataset.
 #' @param Title \href{http://purl.org/dc/elements/1.1/title}{dct:title}, a name given to the resource.
-#' \code{\link{datacite}} allows the use of alternate titles, too. See \code{\link{title}}.
+#' \code{\link{datacite}} allows the use of alternate titles, too. See \code{\link{dataset_title}}.
 #' @param Creator An entity primarily responsible for making the resource. \href{http://purl.org/dc/elements/1.1/creator}{dct:creator}
 #' Corresponds to \code{Creator} in \code{\link{datacite}}. See \code{\link{creator}}.
 #' @param Identifier An unambiguous reference to the resource within a given context.